+++ src/components/ui/popups/PopupContainer.svelte
@@ -12,13 +12,21 @@
 	import TappablePopup from "@/components/ui/popups/tappable/TappablePopup.svelte";
 	import NewPokestopPopup from "@/components/ui/popups/NewPokestopPopup.svelte";
 	import NewPokemonPopup from "@/components/ui/popups/NewPokemonPopup.svelte";
-	import PopupBase, {
+	import PopupBaseDrawer, {
 		type MapObjectPopupProps
-	} from "@/components/ui/popups2/common/PopupBase.svelte";
+	} from "@/components/ui/popups2/common/PopupBaseDrawer.svelte";
 	import { Coords } from "$lib/utils/coordinates";
 	import { watch } from "runed";
 	import * as m from "$lib/paraglide/messages";
 	import { getPopupPropsPokemon } from "@/components/ui/popups2/pokemon/PokemonPopup.svelte";
+	import PopupBaseStatic from "@/components/ui/popups2/common/PopupBaseStatic.svelte";
+	import {fly} from "svelte/transition";
+
+	let {
+		alwaysExpanded = false
+	}: {
+		alwaysExpanded?: boolean
+	} = $props();
 
 	const popupComponents = {
 		[MapObjectType.POKEMON]: NewPokemonPopup,
@@ -51,9 +59,25 @@
 	);
 </script>
 
-<PopupBase
-	open={Boolean(data)}
-	coords={new Coords(snapshotData?.lat ?? 0, snapshotData?.lon ?? 0)}
-	data={snapshotData}
-	{props}
-></PopupBase>
+{#if alwaysExpanded}
+	{#if data}
+		<div
+			class="z-10 w-100 h-screen pointer-events-auto border border-border bg-card/60 backdrop-blur-sm"
+			transition:fly={{ duration: 90, x: 120 }}
+		>
+			<PopupBaseStatic
+				coords={new Coords(snapshotData?.lat ?? 0, snapshotData?.lon ?? 0)}
+				data={snapshotData}
+				{props}
+				onlyShowNavigationButton={false}
+			/>
+		</div>
+	{/if}
+{:else}
+	<PopupBaseDrawer
+		open={Boolean(data)}
+		coords={new Coords(snapshotData?.lat ?? 0, snapshotData?.lon ?? 0)}
+		data={snapshotData}
+		{props}
+	/>
+{/if}
